[fix] 권한 검사 및 작성자 검사 추가

+++ gogildong/src/main/java/com/efub/gogildong/facility/service/FacilityReviewLikeService.java
@@ -2,12 +2,13 @@ package com.efub.gogildong.facility.service;
 
 import com.efub.gogildong.facility.domain.FacilityReview;
 import com.efub.gogildong.facility.domain.FacilityReviewLike;
-import com.efub.gogildong.facility.dto.response.FacilityReviewCommentResponse;
 import com.efub.gogildong.facility.dto.response.FacilityReviewLikeResponse;
 import com.efub.gogildong.facility.respository.FacilityReviewLikeRepository;
-import com.efub.gogildong.facility.respository.FacilityReviewRepository;
 import com.efub.gogildong.global.exception.ExceptionCode;
 import com.efub.gogildong.global.exception.GoGildongException;
+import com.efub.gogildong.global.util.EntityFinder;
+import com.efub.gogildong.schools.domain.School;
+import com.efub.gogildong.schools.service.SchoolViewRequestService;
 import com.efub.gogildong.user.domain.User;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -18,13 +19,18 @@ import org.springframework.transaction.annotation.Transactional;
 public class FacilityReviewLikeService {
 
     private final FacilityReviewLikeRepository facilityReviewLikeRepository;
-    private final FacilityReviewRepository facilityReviewRepository;
+    private final EntityFinder entityFinder;
+    private final SchoolViewRequestService schoolViewRequestService;
 
     // 시설 리뷰 좋아요 생성
     @Transactional
-    public FacilityReviewLikeResponse createFacilityReviewLike(Long reviewId, User user) {
-        FacilityReview review = facilityReviewRepository.findByFacilityReviewId(reviewId)
-                .orElseThrow(() -> new GoGildongException(ExceptionCode.FACILITY_REVIEW_NOT_FOUND));
+    public FacilityReviewLikeResponse createFacilityReviewLike(String loginId, Long reviewId) {
+        User user = entityFinder.getUserByLoginId(loginId);
+        FacilityReview review = entityFinder.getReviewById(reviewId);
+        School school = entityFinder.getSchoolByFacility(review.getFacility());
+
+        // 시설 리뷰 접근 권한 검사
+        schoolViewRequestService.validateViewRequestBySchoolAndUser(school, user);
 
         // 유저가 해당 리뷰에 이미 좋아요를 생성했는지 검사
         if(facilityReviewLikeRepository.existsByFacilityReviewAndUser(review, user)) {
@@ -42,9 +48,22 @@ public class FacilityReviewLikeService {
 
     // 시설 리뷰 좋아요 취소
     @Transactional
-    public void deleteFacilityReviewLike(Long likeId) {
-        FacilityReviewLike reviewLike = facilityReviewLikeRepository.findByFacilityReviewLikeId(likeId)
-                        .orElseThrow(() -> new GoGildongException(ExceptionCode.FACILITY_REVIEW_LIKE_NOT_FOUND));
+    public void deleteFacilityReviewLike(String loginId, Long reviewId) {
+        User user = entityFinder.getUserByLoginId(loginId);
+        FacilityReview review = entityFinder.getReviewById(reviewId);
+        School school = entityFinder.getSchoolByFacility(review.getFacility());
+
+        // 시설 리뷰 접근 권한 검사
+        schoolViewRequestService.validateViewRequestBySchoolAndUser(school, user);
+
+        FacilityReviewLike reviewLike = facilityReviewLikeRepository.findByFacilityReviewAndUser(review, user)
+                .orElseThrow(() -> new GoGildongException(ExceptionCode.FACILITY_REVIEW_LIKE_NOT_FOUND));
+
+        // 좋아요 생성자 검사
+        if(!reviewLike.getUser().getUserId().equals(user.getUserId())) {
+            throw new GoGildongException(ExceptionCode.UNAUTHORIZED_ACCESS);
+        }
+
         facilityReviewLikeRepository.delete(reviewLike);
     }
 }
